modificacion de service y creacion de nueva clase precio

+++ src/main/java/com/HomeRentSolution/ms_reservas/dto/ReservaClienteDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class ReservaClienteDTO {
-    private Integer idReserva;
+    private Long idReserva;
 
     private String tituloPropiedad;
